+++ 3-IoC-DI/ioc-di/src/main/java/com/xevgnov/iocdi/service/FahrenheitTemperatureService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 public class FahrenheitTemperatureService implements TemperatureService {
     
     @Override
-    public long convert(double temperature) {
+    public double convert(double temperature) {
         return (temperature - 32.0) / 1.8;
     }
