+++ 3-IoC-DI/ioc-di/src/main/java/com/xevgnov/iocdi/service/TemperatureService.java
@@ -3,5 +3,5 @@ package com.xevgnov.iocdi.service;
 public interface TemperatureService {
     
     void print(double temperature);
-    long convert(double temperature);
+    double convert(double temperature);
 }
